i changed query function
What changes you have made?
i changed all filter functions and i wrote one bind function with fiter
and i wrote for iin loop

Why this change is necessary?
it removes all if statements

What is the impact of the change?
it easy very easy to read and less complexity